// src/app/shared/interface/country.interface.ts
import { IPaginateModel } from './core.interface';
import { IStates } from './state.interface';

export interface ICountryModel extends IPaginateModel {
  data: ICountry[];
}

export interface ICountry {
  id: number;
  capital: string;
  citizenship: string;
  country_code: string | number;
  currency: string;
  currency_code: string;
  currency_sub_unit: string;
  currency_symbol: string;
  currency_decimals: string | number;
  full_name: string;
  iso_3166_2: string;
  iso_3166_3: string;
  name: string;
  region_code: string | number;
  sub_region_code: string | number;
  eea: string | number;
  calling_code: string | number;
  flag: string;
  state: IStates[];
}